fix: split converter tests by framework dependency
